+++ src/main/java/com/example/sns/post/application/PostImageStore.java
@@ -1,0 +1,9 @@
+package com.example.sns.post.application;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+
+public interface PostImageStore {
+    List<String> savePostImages(List<MultipartFile> images);
+}
